# perf: tweak peak related calculation