--- a/src/helper.rs
+++ b/src/helper.rs
@@ -1,3 +1,4 @@
+use crate::vec;
 use crate::vec::Vec;
 
 pub fn leaf_index_to_pos(index: u64) -> u64 {
@@ -15,75 +16,97 @@ pub fn leaf_index_to_mmr_size(index: u64) -> u64 {
     2 * leaves_count - peak_count
 }
 
-pub fn pos_height_in_tree(mut pos: u64) -> u32 {
-    pos += 1;
-    fn all_ones(num: u64) -> bool {
-        num != 0 && num.count_zeros() == num.leading_zeros()
-    }
-    fn jump_left(pos: u64) -> u64 {
-        let bit_length = 64 - pos.leading_zeros();
-        let most_significant_bits = 1 << (bit_length - 1);
-        pos - (most_significant_bits - 1)
+pub fn pos_height_in_tree(mut pos: u64) -> u8 {
+    if pos == 0 {
+        return 0;
     }
 
-    while !all_ones(pos) {
-        pos = jump_left(pos)
+    let mut peak_size = u64::MAX >> pos.leading_zeros();
+    while peak_size > 0 {
+        if pos >= peak_size {
+            pos -= peak_size;
+        }
+        peak_size >>= 1;
     }
-
-    64 - pos.leading_zeros() - 1
+    pos as u8
 }
 
-pub fn parent_offset(height: u32) -> u64 {
+pub fn parent_offset(height: u8) -> u64 {
     2 << height
 }
 
-pub fn sibling_offset(height: u32) -> u64 {
+pub fn sibling_offset(height: u8) -> u64 {
     (2 << height) - 1
 }
 
-pub fn get_peaks(mmr_size: u64) -> Vec<u64> {
-    let mut pos_s = Vec::new();
-    let (mut height, mut pos) = left_peak_height_pos(mmr_size);
-    pos_s.push(pos);
-    while height > 0 {
-        let peak = match get_right_peak(height, pos, mmr_size) {
-            Some(peak) => peak,
-            None => break,
-        };
-        height = peak.0;
-        pos = peak.1;
-        pos_s.push(pos);
+/// Returns the height of the peaks in the mmr, presented by a bitmap.
+/// for example, for a mmr with 11 leaves, the mmr_size is 19, it will return 0b1010.
+/// 0b1011 indicates that the left peaks are at height 0, 1 and 3.
+///           14
+///        /       \
+///      6          13
+///    /   \       /   \
+///   2     5     9     12     17
+///  / \   /  \  / \   /  \   /  \
+/// 0   1 3   4 7   8 10  11 15  16 18
+///
+/// please note that when the mmr_size is invalid, it will return the bitmap of the last valid mmr.
+/// in the below example, the mmr_size is 6, but it's not a valid mmr, it will return 0b11.
+///   2     5
+///  / \   /  \
+/// 0   1 3   4
+pub fn get_peak_map(mmr_size: u64) -> u64 {
+    if mmr_size == 0 {
+        return 0;
     }
-    pos_s
-}
 
-fn get_right_peak(mut height: u32, mut pos: u64, mmr_size: u64) -> Option<(u32, u64)> {
-    // move to right sibling pos
-    pos += sibling_offset(height);
-    // loop until we find a pos in mmr
-    while pos > mmr_size - 1 {
-        if height == 0 {
-            return None;
+    let mut pos = mmr_size;
+    let mut peak_size = u64::MAX >> pos.leading_zeros();
+    let mut peak_map = 0;
+    while peak_size > 0 {
+        peak_map <<= 1;
+        if pos >= peak_size {
+            pos -= peak_size;
+            peak_map |= 1;
         }
-        // move to left child
-        pos -= parent_offset(height - 1);
-        height -= 1;
+        peak_size >>= 1;
     }
-    Some((height, pos))
+
+    peak_map
 }
 
-fn get_peak_pos_by_height(height: u32) -> u64 {
-    (1 << (height + 1)) - 2
-}
-
-fn left_peak_height_pos(mmr_size: u64) -> (u32, u64) {
-    let mut height = 1;
-    let mut prev_pos = 0;
-    let mut pos = get_peak_pos_by_height(height);
-    while pos < mmr_size {
-        height += 1;
-        prev_pos = pos;
-        pos = get_peak_pos_by_height(height);
+/// Returns the pos of the peaks in the mmr.
+/// for example, for a mmr with 11 leaves, the mmr_size is 19, it will return [14, 17, 18].
+///           14
+///        /       \
+///      6          13
+///    /   \       /   \
+///   2     5     9     12     17
+///  / \   /  \  / \   /  \   /  \
+/// 0   1 3   4 7   8 10  11 15  16 18
+///
+/// please note that when the mmr_size is invalid, it will return the peaks of the last valid mmr.
+/// in the below example, the mmr_size is 6, but it's not a valid mmr, it will return [2, 3].
+///   2     5
+///  / \   /  \
+/// 0   1 3   4
+pub fn get_peaks(mmr_size: u64) -> Vec<u64> {
+    if mmr_size == 0 {
+        return vec![];
     }
-    (height - 1, prev_pos)
+
+    let leading_zeros = mmr_size.leading_zeros();
+    let mut pos = mmr_size;
+    let mut peak_size = u64::MAX >> leading_zeros;
+    let mut peaks = Vec::with_capacity(64 - leading_zeros as usize);
+    let mut peaks_sum = 0;
+    while peak_size > 0 {
+        if pos >= peak_size {
+            pos -= peak_size;
+            peaks.push(peaks_sum + peak_size - 1);
+            peaks_sum += peak_size;
+        }
+        peak_size >>= 1;
+    }
+    peaks
 }