This PR tweaked the peak related calculations by using an unified algorithm, optimized performance and simplified the code. 

Benchmark on my pc shows around 5~6% improvement.

```
MMR insertion/times/10000
                        time:   [8.7440 ms 8.7798 ms 8.8227 ms]
                        change: [-5.5398% -3.7561% -2.1591%] (p = 0.00 < 0.05)

MMR insertion/times/100000
                        time:   [95.951 ms 96.536 ms 97.145 ms]
                        change: [-6.6566% -5.5857% -4.5850%] (p = 0.00 < 0.05)

MMR insertion/times/1000000
                        time:   [1.0498 s 1.0569 s 1.0645 s]
                        change: [-6.7619% -5.8593% -4.9284%] (p = 0.00 < 0.05)

MMR gen proof           time:   [6.6808 µs 6.7964 µs 6.9333 µs]
                        change: [-7.8288% -4.9973% -1.9181%] (p = 0.00 < 0.05)

MMR verify              time:   [6.3648 µs 6.4368 µs 6.5346 µs]
                        change: [-9.4028% -6.0025% -2.4343%] (p = 0.00 < 0.05)
```

please note that some unit test of `get_peaks` fn have been modified since the logic of invalid mmr size calculation is changed, however, it does not affect the valid mmr calculation.