@yangby-cryptape 2023-02-26
In my own humble opinion, it's hard to say whether performance improved or not.

<details><summary>Benchmark on my laptop -- an old mobile workstation.</summary>

<br/>

```diff
  Base Commit: 7a831ab; Turn: 1

- left_index_to_pos           time:   [109.32 ns 111.65 ns 113.38 ns]
+ left_index_to_mmr_size      time:   [15.268 ns 15.299 ns 15.340 ns]
  MMR insertion/times/10000   time:   [8.2108 ms 8.2300 ms 8.2533 ms]
- MMR insertion/times/100000  time:   [100.43 ms 100.61 ms 100.82 ms]
  MMR insertion/times/1000000 time:   [997.36 ms 999.22 ms 1.0015 s]
  MMR gen proof               time:   [6.5195 µs 6.5356 µs 6.5612 µs]
  MMR verify                  time:   [7.0500 µs 7.0688 µs 7.0935 µs]

  Base Commit: 7a831ab; Turn: 2

  left_index_to_pos             time:   [16.717 ns 16.729 ns 16.743 ns]
  left_index_to_mmr_size        time:   [16.245 ns 16.261 ns 16.279 ns]
  MMR insertion/times/10000     time:   [8.2462 ms 8.2552 ms 8.2711 ms]
+ MMR insertion/times/100000    time:   [96.835 ms 96.937 ms 97.042 ms]
- MMR insertion/times/1000000   time:   [1.0624 s 1.0651 s 1.0677 s]
  MMR gen proof                 time:   [6.4087 µs 6.4455 µs 6.4972 µs]
  MMR verify                    time:   [6.8263 µs 6.8387 µs 6.8553 µs]

  PR Commit: c6f69ab; Turn: 1

  left_index_to_pos           time:   [16.791 ns 16.804 ns 16.818 ns]
+ left_index_to_mmr_size      time:   [15.901 ns 15.913 ns 15.926 ns]
+ MMR insertion/times/10000   time:   [7.9936 ms 7.9999 ms 8.0069 ms]
  MMR insertion/times/100000  time:   [97.007 ms 97.149 ms 97.322 ms]
  MMR insertion/times/1000000 time:   [1.0148 s 1.0166 s 1.0185 s]
  MMR gen proof               time:   [6.4738 µs 6.4880 µs 6.5103 µs]
- MMR verify                  time:   [7.3846 µs 7.4172 µs 7.4816 µs]

  PR Commit: c6f69ab; Turn: 2

+ left_index_to_pos           time:   [15.758 ns 15.776 ns 15.798 ns]
+ left_index_to_mmr_size      time:   [15.307 ns 15.325 ns 15.345 ns]
- MMR insertion/times/10000   time:   [8.4411 ms 8.4483 ms 8.4560 ms]
  MMR insertion/times/100000  time:   [97.629 ms 97.707 ms 97.790 ms]
+ MMR insertion/times/1000000 time:   [967.80 ms 969.13 ms 970.56 ms]
  MMR gen proof               time:   [6.4337 µs 6.4567 µs 6.4972 µs]
  MMR verify                  time:   [6.8034 µs 6.8227 µs 6.8576 µs]

  PR Commit: c6f69ab; Turn: 3

- left_index_to_pos           time:   [18.596 ns 18.641 ns 18.694 ns]
- left_index_to_mmr_size      time:   [18.147 ns 18.164 ns 18.184 ns]
  MMR insertion/times/10000   time:   [8.1714 ms 8.1931 ms 8.2214 ms]
  MMR insertion/times/100000  time:   [99.261 ms 99.389 ms 99.534 ms]
  MMR insertion/times/1000000 time:   [1.0185 s 1.1325 s 1.3545 s]
  MMR gen proof               time:   [6.5246 µs 6.5456 µs 6.5814 µs]
  MMR verify                  time:   [6.8794 µs 6.9484 µs 7.0379 µs]
```

</details>

The result seems quite randomized.